pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-mosaic"
version = "1.0.0"
description = "Generate Richter-style pixel mosaics: arrange a fixed palette of random colors to approximate any image using optimal assignment algorithms."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Max"},
]
keywords = ["pixel-art", "mosaic", "optimization", "hungarian-algorithm", "simulated-annealing", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "scipy>=1.11",
    "scikit-image>=0.21",
    "typer[all]>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.5",
]
demo = [
    "streamlit>=1.28",
]
all = ["pixel-mosaic[dev,demo]"]

[project.scripts]
pixel-mosaic = "pixel_mosaic.cli:app"

[project.urls]
Repository = "https://github.com/YOURNAME/pixel-mosaic"

[tool.setuptools.packages.find]
include = ["pixel_mosaic*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["B008"]  # typer uses function calls in defaults by design

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
